Oops...

+++ Ryujinx/Config.cs
@@ -60,20 +60,20 @@ namespace Ryujinx
 
             XInput = new XInputController
             {
-                ButtonA     = Convert.ToInt16(Parser.Value("Controls_XInput_Button_A")),
-                ButtonB     = Convert.ToInt16(Parser.Value("Controls_XInput_Button_B")),
-                ButtonX     = Convert.ToInt16(Parser.Value("Controls_XInput_Button_X")),
-                ButtonY     = Convert.ToInt16(Parser.Value("Controls_XInput_Button_Y")),
-                ButtonPlus  = Convert.ToInt16(Parser.Value("Controls_XInput_Button_Plus")),
-                ButtonMinus = Convert.ToInt16(Parser.Value("Controls_XInput_Button_Minus")),
-                DPadUp      = Convert.ToInt16(Parser.Value("Controls_XInput_DPad_Up")),
-                DPadDown    = Convert.ToInt16(Parser.Value("Controls_XInput_DPad_Down")),
-                DPadLeft    = Convert.ToInt16(Parser.Value("Controls_XInput_DPad_Left")),
-                DPadRight   = Convert.ToInt16(Parser.Value("Controls_XInput_DPad_Right")),
-                ButtonL     = Convert.ToInt16(Parser.Value("Controls_XInput_Button_L")),
-                ButtonZL    = Convert.ToInt16(Parser.Value("Controls_XInput_Button_ZL")),
-                ButtonR     = Convert.ToInt16(Parser.Value("Controls_XInput_Button_R")),
-                ButtonZR    = Convert.ToInt16(Parser.Value("Controls_XInput_Button_ZR"))
+                ButtonA     = Convert.ToInt32(Parser.Value("Controls_XInput_Button_A")),
+                ButtonB     = Convert.ToInt32(Parser.Value("Controls_XInput_Button_B")),
+                ButtonX     = Convert.ToInt32(Parser.Value("Controls_XInput_Button_X")),
+                ButtonY     = Convert.ToInt32(Parser.Value("Controls_XInput_Button_Y")),
+                ButtonPlus  = Convert.ToInt32(Parser.Value("Controls_XInput_Button_Plus")),
+                ButtonMinus = Convert.ToInt32(Parser.Value("Controls_XInput_Button_Minus")),
+                DPadUp      = Convert.ToInt32(Parser.Value("Controls_XInput_DPad_Up")),
+                DPadDown    = Convert.ToInt32(Parser.Value("Controls_XInput_DPad_Down")),
+                DPadLeft    = Convert.ToInt32(Parser.Value("Controls_XInput_DPad_Left")),
+                DPadRight   = Convert.ToInt32(Parser.Value("Controls_XInput_DPad_Right")),
+                ButtonL     = Convert.ToInt32(Parser.Value("Controls_XInput_Button_L")),
+                ButtonZL    = Convert.ToInt32(Parser.Value("Controls_XInput_Button_ZL")),
+                ButtonR     = Convert.ToInt32(Parser.Value("Controls_XInput_Button_R")),
+                ButtonZR    = Convert.ToInt32(Parser.Value("Controls_XInput_Button_ZR"))
             };
 
             FakeJoyCon = new JoyCon
